requirements.txt
# --- 核心 ---
langchain-core>=1.3
langchain>=1.2
langchain-openai>=1.2
langchain-deepseek>=1.0
langchain-community>=0.3
langchain-chroma>=0.2
langchain-text>=0.1

# --- 向量/检索 ---
chromadb>=0.6
jieba>=0.42
rank-bm25>=0.2

# --- Rerank（远程 API，可选）---
httpx>=0.27

# --- 本地 Embedding（可选，EMBEDDING_PROVIDER=local 时需要）---
# langchain-huggingface>=0.1
# sentence-transformers>=3.0

# --- 服务 ---
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.0

# --- 工具 ---
python-dotenv>=1.0
PyYAML>=6.0
